Preview aded to the archive_box
if you use archive with class=box you may want to have a preview of
severeal signs. Use parameter p_sign to specify a number (e.g. 200) to
be displayed as a preview. Further you may use a little simple style
within such a preview. then set cws=1. In the event you want to see the
original styling the set cws=2. If cws is missing then only the text
will be displayed.